+++ example/src/main/java/example/ExampleCsvWriterWithComments.java
@@ -1,7 +1,5 @@
 package example;
 
-import java.io.StringWriter;
-
 import de.siegmar.fastcsv.writer.CsvWriter;
 
 /**
@@ -10,13 +8,10 @@ import de.siegmar.fastcsv.writer.CsvWriter;
 public class ExampleCsvWriterWithComments {
 
     public static void main(final String[] args) {
-        final StringWriter sw = new StringWriter();
-        CsvWriter.builder().build(sw)
+        CsvWriter.builder().toConsole()
             .writeComment("A comment can be placed\nanywhere")
             .writeRecord("field 1", "field 2", "field 3\n#with a line break")
             .writeComment("in the CSV file");
-
-        System.out.println(sw);
     }
 
 }