+++ example/src/main/java/example/ExampleCsvWriterWithNonStandardControlCharacters.java
@@ -1,7 +1,5 @@
 package example;
 
-import java.io.StringWriter;
-
 import de.siegmar.fastcsv.writer.CsvWriter;
 import de.siegmar.fastcsv.writer.LineDelimiter;
 
@@ -11,8 +9,6 @@ import de.siegmar.fastcsv.writer.LineDelimiter;
 public class ExampleCsvWriterWithNonStandardControlCharacters {
 
     public static void main(final String[] args) {
-        final StringWriter sw = new StringWriter();
-
         // The default configuration uses a comma as field separator,
         // a double quote as quote character and
         // a CRLF as line delimiter.
@@ -20,11 +16,9 @@ public class ExampleCsvWriterWithNonStandardControlCharacters {
             .fieldSeparator(';')
             .quoteCharacter('\'')
             .lineDelimiter(LineDelimiter.LF)
-            .build(sw)
+            .toConsole()
             .writeRecord("header1", "header2")
             .writeRecord("value1", "value;2");
-
-        System.out.println(sw);
     }
 
 }