+++ lib/src/test/java/de/siegmar/fastcsv/writer/FastBufferedWriterTest.java
@@ -4,13 +4,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.IOException;
 import java.io.StringWriter;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import org.junit.jupiter.api.Test;
 
 class FastBufferedWriterTest {
 
     private final StringWriter sw = new StringWriter();
-    private final CsvWriter.FastBufferedWriter cw = new CsvWriter.FastBufferedWriter(sw, 8192);
+    private final FastBufferedWriter cw = new FastBufferedWriter(sw, 8192, false, false);
 
     @Test
     void appendSingle() throws IOException {
@@ -56,11 +57,57 @@ class FastBufferedWriterTest {
     }
 
     @Test
-    void appendLarge() throws IOException {
+    void appendLargeString() throws IOException {
         final String sb = buildLargeData();
         cw.write(sb, 0, sb.length());
 
         assertThat(sw).asString().isEqualTo(sb);
+    }
+
+    @Test
+    void appendLargeArray() throws IOException {
+        final char[] data = buildLargeData().toCharArray();
+        cw.write(data, 0, data.length);
+
+        assertThat(sw).asString().isEqualTo(new String(data));
+    }
+
+    @Test
+    void autoFlushBuffer() throws IOException {
+        final var stringWriter = new StringWriter() {
+            @Override
+            public void flush() {
+                throw new UnsupportedOperationException();
+            }
+        };
+
+        final var fbw = new FastBufferedWriter(stringWriter, 8, true, false);
+
+        fbw.write("foo");
+        assertThat(stringWriter).asString().isEmpty();
+
+        fbw.endRecord();
+        assertThat(stringWriter).asString().isEqualTo("foo");
+    }
+
+    @Test
+    void autoFlushWriter() throws IOException {
+        final AtomicInteger flushCount = new AtomicInteger();
+        final var stringWriter = new StringWriter() {
+            @Override
+            public void flush() {
+                flushCount.incrementAndGet();
+            }
+        };
+
+        final var fbw = new FastBufferedWriter(stringWriter, 8, true, true);
+
+        fbw.write("foo");
+        assertThat(stringWriter).asString().isEmpty();
+
+        fbw.endRecord();
+        assertThat(stringWriter).asString().isEqualTo("foo");
+        assertThat(flushCount).hasValue(1);
     }
 
     private String buildLargeData() {